[GLUTEN-6705]][CORE][VL][CH] [Part-2] Rework CumnarWriteFilesExec
